pyproject.toml
[tool.poetry]
name = "discount-alert-bot"
version = "0.1.0"
description = "Telegram chatbot that notifies users about discounts on products from their wish lists."
authors = ["Ihar Yazerski"]
readme = "README.md"
license = "MIT"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
sqlalchemy = "^2.0.36"
loguru = "^0.7.2"
alembic = "^1.14.0"
pydantic = "^2.9.2"
pydantic-settings = "^2.6.1"
celery = "^5.4.0"
psycopg2-binary = "^2.9.10"

[tool.poetry.group.dev.dependencies]
ruff = "^0.7.4"
ipython = "^8.29.0"
pre-commit = "^4.0.1"
mypy = "^1.13.0"
isort = "^5.13.2"

[tool.poetry.group.bot_interface.dependencies]
python-telegram-bot = "^21.7"

[tool.poetry.group.bot_engine.dependencies]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120
indent-width = 4

[tool.ruff.format]
quote-style = "double"

[tool.ruff.lint.isort]
combine-as-imports = true
